% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{temp_load}
\alias{temp_load}
\title{temp_load}
\usage{
temp_load(version_id, format = "rds", fpath = NA)
}
\arguments{
\item{version_id}{character database version string}

\item{format}{character choice of rds or sqlite}

\item{fpath}{file.path optionally specify custom location of rds file}
}
\description{
Load files from local file system
}
\examples{
\dontrun{
dt  <- temp_load("1")
dt2 <- temp_load("2")
}
}
