% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{temp_ls}
\alias{temp_ls}
\title{temp_ls}
\usage{
temp_ls(version_id, ...)
}
\arguments{
\item{version_id}{character version id}

\item{...}{extra arguments passed to list.files}
}
\description{
temp_ls
}
\examples{
\dontrun{
temp_ls("1")
}
}
