% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{temp_get}
\alias{temp_get}
\title{temp_get}
\usage{
temp_get(dest_folder = NA)
}
\arguments{
\item{dest_folder}{file.path optional will default to the location returned by \code{\link[rappdirs]{user_data_dir}}.}
}
\description{
Retrieves external files and store in file cache.
}
\examples{
\dontrun{
temp_get()
}
}
