% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{temp_ingest}
\alias{temp_ingest}
\title{Ingest flat files}
\usage{
temp_ingest(version_id, folder = NA, skip = NA)
}
\arguments{
\item{version_id}{character temp database version string}

\item{folder}{file.path to data folder. optional.}

\item{skip}{numeric vector of lines to skip on file read. optional.}
}
\description{
Ingest data from component flat files
}
\examples{
\dontrun{
temp_ingest("1")
}

}
